+++ core/src/db/snapshot/Resources.h
@@ -328,6 +328,25 @@ class SizeField {
     SIZE_TYPE size_;
 };
 
+class RowCountField {
+ public:
+    explicit RowCountField(SIZE_TYPE size) : size_(size) {
+    }
+
+    SIZE_TYPE
+    GetRowCount() const {
+        return size_;
+    }
+
+    void
+    SetRowCount(SIZE_TYPE row_cnt) {
+        size_ = row_cnt;
+    }
+
+ protected:
+    SIZE_TYPE size_;
+};
+
 ///////////////////////////////////////////////////////////////////////////////
 
 class Collection : public BaseResource,
@@ -356,6 +375,7 @@ class CollectionCommit : public BaseResource,
                          public CollectionIdField,
                          public SchemaIdField,
                          public MappingsField,
+                         public RowCountField,
                          public SizeField,
                          public IdField,
                          public LsnField,
@@ -368,8 +388,8 @@ class CollectionCommit : public BaseResource,
     using MapT = std::map<ID_TYPE, Ptr>;
     using ScopedMapT = std::map<ID_TYPE, ScopedResource<CollectionCommit>>;
     using VecT = std::vector<Ptr>;
-    CollectionCommit(ID_TYPE collection_id, ID_TYPE schema_id, const MappingT& mappings = {}, SIZE_TYPE size = 0,
-                     ID_TYPE id = 0, LSN_TYPE lsn = 0, State status = PENDING,
+    CollectionCommit(ID_TYPE collection_id, ID_TYPE schema_id, const MappingT& mappings = {}, SIZE_TYPE row_cnt = 0,
+                     SIZE_TYPE size = 0, ID_TYPE id = 0, LSN_TYPE lsn = 0, State status = PENDING,
                      TS_TYPE created_on = GetMicroSecTimeStamp(), TS_TYPE UpdatedOnField = GetMicroSecTimeStamp());
 };
 
@@ -402,6 +422,7 @@ class PartitionCommit : public BaseResource,
                         public CollectionIdField,
                         public PartitionIdField,
                         public MappingsField,
+                        public RowCountField,
                         public SizeField,
                         public IdField,
                         public LsnField,
@@ -415,8 +436,8 @@ class PartitionCommit : public BaseResource,
     using VecT = std::vector<Ptr>;
     static constexpr const char* Name = "PartitionCommit";
 
-    PartitionCommit(ID_TYPE collection_id, ID_TYPE partition_id, const MappingT& mappings = {}, SIZE_TYPE size = 0,
-                    ID_TYPE id = 0, LSN_TYPE lsn = 0, State status = PENDING,
+    PartitionCommit(ID_TYPE collection_id, ID_TYPE partition_id, const MappingT& mappings = {}, SIZE_TYPE row_cnt = 0,
+                    SIZE_TYPE size = 0, ID_TYPE id = 0, LSN_TYPE lsn = 0, State status = PENDING,
                     TS_TYPE created_on = GetMicroSecTimeStamp(), TS_TYPE UpdatedOnField = GetMicroSecTimeStamp());
 
     std::string
@@ -458,6 +479,7 @@ class SegmentCommit : public BaseResource,
                       public PartitionIdField,
                       public SegmentIdField,
                       public MappingsField,
+                      public RowCountField,
                       public SizeField,
                       public IdField,
                       public LsnField,
@@ -472,7 +494,7 @@ class SegmentCommit : public BaseResource,
     static constexpr const char* Name = "SegmentCommit";
 
     SegmentCommit(ID_TYPE schema_id, ID_TYPE partition_id, ID_TYPE segment_id, const MappingT& mappings = {},
-                  SIZE_TYPE size = 0, ID_TYPE id = 0, LSN_TYPE lsn = 0, State status = PENDING,
+                  SIZE_TYPE row_cnt = 0, SIZE_TYPE size = 0, ID_TYPE id = 0, LSN_TYPE lsn = 0, State status = PENDING,
                   TS_TYPE created_on = GetMicroSecTimeStamp(), TS_TYPE UpdatedOnField = GetMicroSecTimeStamp());
 
     std::string
@@ -488,6 +510,7 @@ class SegmentFile : public BaseResource,
                     public PartitionIdField,
                     public SegmentIdField,
                     public FieldElementIdField,
+                    public RowCountField,
                     public SizeField,
                     public IdField,
                     public LsnField,
@@ -502,7 +525,7 @@ class SegmentFile : public BaseResource,
     static constexpr const char* Name = "SegmentFile";
 
     SegmentFile(ID_TYPE collection_id, ID_TYPE partition_id, ID_TYPE segment_id, ID_TYPE field_element_id,
-                SIZE_TYPE size = 0, ID_TYPE id = 0, LSN_TYPE lsn = 0, State status = PENDING,
+                SIZE_TYPE row_cnt = 0, SIZE_TYPE size = 0, ID_TYPE id = 0, LSN_TYPE lsn = 0, State status = PENDING,
                 TS_TYPE created_on = GetMicroSecTimeStamp(), TS_TYPE UpdatedOnField = GetMicroSecTimeStamp());
 };