+++ core/src/index/knowhere/knowhere/index/vector_offset_index/IndexIVF_NM.h
@@ -59,14 +59,10 @@ class IVF_NM : public VecIndex, public OffsetBaseIndex {
 #endif
 
     int64_t
-    Count() override {
-        return index_->ntotal;
-    }
+    Count() override;
 
     int64_t
-    Dim() override {
-        return index_->d;
-    }
+    Dim() override;
 
 #if 0
     DatasetPtr