+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Please mark all changes in change log and use the issue from GitHub
 -   \#2598 Fix Milvus docker image report illegal instruction
 -   \#2617 Fix HNSW and RNSG index files size
 -   \#2637 Suit the range of HNSW parameters
+-   \#2642 Create index failed and server crashed
 -   \#2649 Search parameter of annoy has conflict with document
 -   \#2690 Remove body parser in show-partitions endpoints